+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [Unreleased]
+
+### Changed
+
+- concord-server: fixed an issue preventing the process wait
+condition history from being correctly filled in.
+
+
+
 ## [1.32.0] - 2019-09-04
 
 ### Added